Extract continuations for Async
